refactor(server-libs): extract models/usage/health route handlers — no inline route bodies remain
